refactor(ast): forward exports from node files and combine build/finalize
factory.ts is no longer a re-export barrel: index.ts forwards create*/is*
straight from each node file, and factory.ts keeps only createFile and update.
Internal modules and tests import create* directly from the node files, which
also removes the old factory <-> utils/ast import cycle.

defineNode's finalize is now a boolean: when set, the registry reuses create as
the finalizer instead of a separate function, so build and finalize share one
implementation and the self-referencing def annotations are gone.
