+++ packages/ast/src/index.ts
@@ -1,33 +1,7 @@
 export { httpMethods, schemaTypes } from './constants.ts'
 export { applyDedupe, buildDedupePlan } from './dedupe.ts'
 export { defineSchemaDialect } from './dialect.ts'
-export {
-  createArrowFunction,
-  createBreak,
-  createConst,
-  createExport,
-  createFile,
-  createFunction,
-  createFunctionParameter,
-  createFunctionParameters,
-  createImport,
-  createInput,
-  createStreamInput,
-  createJsx,
-  createOperation,
-  createOutput,
-  createParameter,
-  createParameterGroup,
-  createParamsType,
-  createProperty,
-  createResponse,
-  createSchema,
-  createSource,
-  createText,
-  createType,
-  syncOptionality,
-  update,
-} from './factory.ts'
+export { createFile, update } from './factory.ts'
 export {
   isArrowFunctionNode,
   isBreakNode,
@@ -58,18 +32,43 @@ export {
 } from './guards.ts'
 export { defineNode } from './node.ts'
 export type { NodeDef } from './node.ts'
-export { arrowFunctionDef, breakDef, constDef, functionDef, jsxDef, textDef, typeDef } from './nodes/code.ts'
+export { syncOptionality } from './node.ts'
+export {
+  arrowFunctionDef,
+  breakDef,
+  constDef,
+  createArrowFunction,
+  createBreak,
+  createConst,
+  createFunction,
+  createJsx,
+  createText,
+  createType,
+  functionDef,
+  jsxDef,
+  textDef,
+  typeDef,
+} from './nodes/code.ts'
 export { contentDef } from './nodes/content.ts'
-export { exportDef, fileDef, importDef, sourceDef } from './nodes/file.ts'
-export { functionParameterDef, functionParametersDef, parameterGroupDef, paramsTypeDef } from './nodes/function.ts'
-export { inputDef } from './nodes/input.ts'
-export { operationDef } from './nodes/operation.ts'
-export { outputDef } from './nodes/output.ts'
-export { parameterDef } from './nodes/parameter.ts'
-export { propertyDef } from './nodes/property.ts'
+export { createExport, createImport, createSource, exportDef, fileDef, importDef, sourceDef } from './nodes/file.ts'
+export {
+  createFunctionParameter,
+  createFunctionParameters,
+  createParameterGroup,
+  createParamsType,
+  functionParameterDef,
+  functionParametersDef,
+  parameterGroupDef,
+  paramsTypeDef,
+} from './nodes/function.ts'
+export { createInput, createStreamInput, inputDef } from './nodes/input.ts'
+export { createOperation, operationDef } from './nodes/operation.ts'
+export { createOutput, outputDef } from './nodes/output.ts'
+export { createParameter, parameterDef } from './nodes/parameter.ts'
+export { createProperty, propertyDef } from './nodes/property.ts'
 export { requestBodyDef } from './nodes/requestBody.ts'
-export { responseDef } from './nodes/response.ts'
-export { schemaDef } from './nodes/schema.ts'
+export { createResponse, responseDef } from './nodes/response.ts'
+export { createSchema, schemaDef } from './nodes/schema.ts'
 export { createPrinterFactory, definePrinter } from './printer.ts'
 export { signatureOf } from './signature.ts'
 export { mergeAdjacentObjectsLazy, setDiscriminatorEnum, setEnumName, simplifyUnion } from './transformers.ts'
