+++ packages/ast/src/printer.test.ts
@@ -1,5 +1,6 @@
 import { describe, expect, expectTypeOf, it } from 'vitest'
-import { createProperty, createSchema } from './factory.ts'
+import { createProperty } from './nodes/property.ts'
+import { createSchema } from './nodes/schema.ts'
 import type { PrinterFactoryOptions } from './printer.ts'
 import { definePrinter } from './printer.ts'
