+++ packages/ast/src/visitor.test.ts
@@ -1,6 +1,8 @@
 import { describe, expect, expectTypeOf, it } from 'vitest'
-import { createInput, createProperty, createSchema } from './factory.ts'
 import { buildSampleTree } from './mocks.ts'
+import { createInput } from './nodes/input.ts'
+import { createProperty } from './nodes/property.ts'
+import { createSchema } from './nodes/schema.ts'
 import type { ContentNode } from './nodes/content.ts'
 import type { InputNode } from './nodes/input.ts'
 import type { OperationNode } from './nodes/operation.ts'
